feat(front): new trinity main page
