--- components/layout/FooterLayout.vue
@@ -1,18 +1,0 @@
-<template>
-  <div class="flex w-full h-full bg-stone-900 place-items-center justify-center bg-opacity-90">
-    <div class="text-base text-stone-600">TRINITY / 2023 / Powered by Devjk. All Rights Reserved.
-    </div>
-  </div>
-</template>
-
-<script lang="ts">
-import {defineComponent} from 'vue'
-
-export default defineComponent({
-  name: "FooterLayout"
-})
-</script>
-
-<style scoped>
-
-</style>
