+++ components/layout/default/FooterLayout.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="flex w-full h-full bg-stone-900 place-items-center justify-center bg-opacity-50 backdrop-blur">
+    <div class="text-base text-stone-600">TRINITY / 2023 / Powered by Devjk. All Rights Reserved.
+    </div>
+  </div>
+</template>
+
+<script lang="ts">
+import {defineComponent} from 'vue'
+
+export default defineComponent({
+  name: "FooterLayout"
+})
+</script>
+
+<style scoped>
+
+</style>
