+++ components/layout/default/HeaderLayout.vue
@@ -1,0 +1,25 @@
+<template>
+  <div class="flex flex-col w-full h-full bg-stone-800 bg-opacity-50 backdrop-blur">
+    <div class="flex m-5 mt-20 justify-center">
+      <img class="object-cover" src="/favicon-192x192.png"/>
+    </div>
+    <div class="flex flex-col m-5 mb-10 items-center">
+      <div class="tracking-wide font-extrabold subpixel-antialiased text-9xl text-stone-50">
+        TRINITY
+      </div>
+      <div class="tracking-wide font-extrabold subpixel-antialiased text-3xl text-stone-300">
+        The Best solutions for your personal branding.
+      </div>
+    </div>
+
+
+  </div>
+</template>
+
+<script setup lang="ts">
+
+</script>
+
+<style scoped>
+
+</style>
