+++ pages/[userName]/about.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="flex flex-col bg-stone-800 h-full p-3 select-auto">
+  <div class="flex flex-col bg-stone-800 h-full p-3 select-auto bg-opacity-50 backdrop-blur">
     <div class="flex w-full h-full justify-center place-items-center text-lime-200 text-6xl">
       WE'RE PREPARING THIS PAGE
       THIS IS ABOUT PAGE
